+++ packages/terra-date-picker/CHANGELOG.md
@@ -7,6 +7,9 @@ Unreleased
 * Limit the supported dates to between 01/01/1900 and 12/31/2100 because the year dropdown only contains years between 1900 and 2100.
 * Disable the Today button if today is a disabled day. (i.e.; outside the `minDate`/`maxDate` range, in `excludeDates`, not in `includeDates`, or filtered by `filterDate`)
 
+### Changed
+* Apply themes by context and directly include packaged themes.
+
 4.37.0 - (May 19, 2020)
 ------------------
 ### Changed